+++ 340-ce-nextgen/src/main/java/io/harness/ccm/remote/resources/BusinessMappingResource.java
@@ -16,6 +16,7 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.ccm.audittrails.events.CostCategoryCreateEvent;
 import io.harness.ccm.audittrails.events.CostCategoryDeleteEvent;
 import io.harness.ccm.audittrails.events.CostCategoryUpdateEvent;
+import io.harness.ccm.rbac.CCMRbacHelper;
 import io.harness.ccm.views.businessMapping.entities.BusinessMapping;
 import io.harness.ccm.views.businessMapping.service.intf.BusinessMappingService;
 import io.harness.outbox.api.OutboxService;
@@ -55,6 +56,7 @@ public class BusinessMappingResource {
   @Inject BusinessMappingService businessMappingService;
   @Inject @Named(OUTBOX_TRANSACTION_TEMPLATE) private TransactionTemplate transactionTemplate;
   @Inject private OutboxService outboxService;
+  @Inject CCMRbacHelper rbacHelper;
 
   private final RetryPolicy<Object> transactionRetryPolicy = DEFAULT_TRANSACTION_RETRY_POLICY;
 
@@ -64,6 +66,7 @@ public class BusinessMappingResource {
   @ApiOperation(value = "Create Business Mapping", nickname = "createBusinessMapping")
   public RestResponse<BusinessMapping> save(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId, BusinessMapping businessMapping) {
+    rbacHelper.checkCostCategoryEditPermission(accountId, null, null);
     BusinessMapping costCategory = businessMappingService.save(businessMapping);
     Failsafe.with(transactionRetryPolicy).get(() -> transactionTemplate.execute(status -> {
       outboxService.save(new CostCategoryCreateEvent(accountId, costCategory.toDTO()));
@@ -77,6 +80,7 @@ public class BusinessMappingResource {
   @ExceptionMetered
   @ApiOperation(value = "Get List Of Business Mappings", nickname = "getBusinessMappingList")
   public RestResponse<List<BusinessMapping>> list(@QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId) {
+    rbacHelper.checkCostCategoryViewPermission(accountId, null, null);
     return new RestResponse<>(businessMappingService.list(accountId));
   }
 
@@ -87,6 +91,7 @@ public class BusinessMappingResource {
   @ApiOperation(value = "Get Business Mapping", nickname = "getBusinessMapping")
   public RestResponse<BusinessMapping> get(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId, @PathParam("id") String businessMappingId) {
+    rbacHelper.checkCostCategoryViewPermission(accountId, null, null);
     return new RestResponse<>(businessMappingService.get(businessMappingId, accountId));
   }
 
@@ -96,6 +101,7 @@ public class BusinessMappingResource {
   @ApiOperation(value = "Update Business Mapping", nickname = "updateBusinessMapping")
   public RestResponse<String> update(
       @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId, BusinessMapping businessMapping) {
+    rbacHelper.checkCostCategoryEditPermission(accountId, null, null);
     BusinessMapping oldCostCategory = businessMappingService.get(businessMapping.getUuid(), accountId);
     BusinessMapping newCostCategory = businessMappingService.update(businessMapping);
     Failsafe.with(transactionRetryPolicy).get(() -> transactionTemplate.execute(status -> {
@@ -112,6 +118,7 @@ public class BusinessMappingResource {
   @ApiOperation(value = "Delete Business Mapping", nickname = "deleteBusinessMapping")
   public RestResponse<String> delete(@NotEmpty @QueryParam(NGCommonEntityConstants.ACCOUNT_KEY) String accountId,
       @PathParam("id") String businessMappingId) {
+    rbacHelper.checkCostCategoryDeletePermission(accountId, null, null);
     BusinessMapping costCategory = businessMappingService.get(businessMappingId, accountId);
     businessMappingService.delete(businessMappingId, accountId);
     Failsafe.with(transactionRetryPolicy).get(() -> transactionTemplate.execute(status -> {
